1-memcpy.c

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -6,15 +6,14 @@
 #include "main.h"
 
 /**
- * _memcpy - Copies @n bytes from the memory area pointed
- * to by @src into that pointed to by @dest.
- * @dest: A pointer to the memory area to copy @src into.
- * @src: The source buffer to copy characters from.
- * @n: The number of bytes to copy from @src.
+ * char *_memcpy - Copies n bytes from memory area src to memory area dest
+ * @dest: destination
+ * @src: source
+ * @n: number of bytes
  *
- * Return: A pointer to the destination buffer @dest.
+ * Return: A pointer to the destination buffer
  */
-char *_memcpy(char *dest, char *src, size_t n)
+char *_memcpy(char *dest, char *src, unsigned int n)
 {
 	unsigned int index;
 	unsigned char *destination = dest;
